+++ HISTORY.rst
@@ -9,7 +9,7 @@ History
   attribute-keys which can be passed to tag-renderer.
   [thet, 2012-12-05]
 
-- Allow custom attributes on textarea_renderer.
+- Allow custom attributes on textarea_renderer and select_edit_renderer.
   [thet, 2012-12-05]
 
 2.0.2